# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "API documentation for the application",
        "version": "1.0.0"
    },
    "paths": {
        "/api/master/rekening/mutasi": {
            "get": {
                "tags": [
                    "Mutasi"
                ],
                "summary": "Get data mutasi",
                "operationId": "2e65cba4c6347f5690914f24ff952d98",
                "responses": {
                    "200": {
                        "description": "Data mutasi berhasil ditemukan"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/master/rekening/{rekening_id}/mutasi": {
            "get": {
                "tags": [
                    "Mutasi"
                ],
                "summary": "Get data mutasi rekening",
                "operationId": "ea4e2cae70c52db374caaedcdb993ceb",
                "parameters": [
                    {
                        "name": "rekening_id",
                        "in": "path",
                        "description": "ID rekening",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data mutasi rekening berhasil ditemukan"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/master/rekening": {
            "get": {
                "tags": [
                    "Rekening"
                ],
                "summary": "Get data rekening",
                "operationId": "0592873c51fc82a73cb65321b6fb2958",
                "responses": {
                    "200": {
                        "description": "Data rekening berhasil ditemukan"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/master/rekening/{id}": {
            "get": {
                "tags": [
                    "Rekening"
                ],
                "summary": "Get data rekening",
                "operationId": "e9244f38c29157aee5efe2ba40d7949c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID rekening",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data rekening berhasil ditemukan"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register new user",
                "operationId": "a97e586990d0776b795fe45dfced8147",
                "requestBody": {
                    "description": "Data that will be used to register new user",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "username",
                                    "email",
                                    "name",
                                    "password"
                                ],
                                "properties": {
                                    "username": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "user123"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@gmail.com"
                                    },
                                    "name": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "User 123"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success register new user"
                    }
                }
            }
        },
        "/api/laporan/lainlain": {
            "get": {
                "tags": [
                    "Servis Kategori Lain-lain"
                ],
                "summary": "Get data servis tipe lain-lain",
                "operationId": "33f7f3383af967bf838fe41e97f330c6",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Kata kunci untuk mencari data servis tipe lain-lain",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data servis tipe lain-lain berhasil ditemukan"
                    }
                }
            },
            "post": {
                "tags": [
                    "Servis Kategori Lain-lain"
                ],
                "summary": "Tambah data servis tipe lain-lain",
                "operationId": "38f8d303e851d62169d2caf7fac1295b",
                "requestBody": {
                    "description": "Data yang dibutuhkan untuk menambah data servis tipe lain-lain",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "master_armada_id",
                                    "nomor_nota",
                                    "nama_toko",
                                    "tanggal_servis",
                                    "nota_beli_items",
                                    "kategori_servis",
                                    "nama_tujuan_lain",
                                    "keterangan_lain",
                                    "nominal_lain",
                                    "jumlah_lain",
                                    "total_lain"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data servis berhasil ditambahkan"
                    },
                    "400": {
                        "description": "Data servis gagal ditambahkan"
                    },
                    "422": {
                        "description": "Data yang dibutuhkan tidak lengkap"
                    }
                }
            }
        },
        "/api/laporan/lainlain/{id}": {
            "get": {
                "tags": [
                    "Servis Kategori Lain-lain"
                ],
                "summary": "Get data servis tipe lain-lain by id",
                "operationId": "2284667c4991afb85105f922eb534900",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID dari data servis tipe lain-lain",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data servis tipe lain-lain berhasil ditemukan"
                    }
                }
            }
        },
        "/api/laporan/servis": {
            "get": {
                "tags": [
                    "Servis Kategori Servis"
                ],
                "summary": "Get data servis tipe servis",
                "operationId": "dfe19fc4c76814af5f3f53d2e511b090",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Kata kunci untuk mencari data servis tipe servis",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data servis tipe servis berhasil ditemukan"
                    }
                }
            }
        },
        "/api/laporan/servis/mutasi": {
            "post": {
                "tags": [
                    "Servis Mutasi"
                ],
                "summary": "Create Servis Mutasi",
                "operationId": "acc1544e453a0d6326f56cfcbb671b59",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "servis_id",
                                    "master_rekening_id",
                                    "nominal"
                                ],
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Servis mutasi berhasil dibuat"
                    },
                    "422": {
                        "description": "Data tidak valid"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login user",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "username",
                                    "password"
                                ],
                                "properties": {
                                    "username": {
                                        "type": "string",
                                        "example": "boss"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "Abc12345"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User berhasil login"
                    },
                    "401": {
                        "description": "Username dan Password anda salah"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}